+++ tutorial/shared/src/androidMain/kotlin/com/example/tutorial/shared/auth/CredentialsManager.kt
@@ -1,9 +1,9 @@
-package com.example.tutorial.androidApp
+package com.example.tutorial.shared.auth
 
 import android.content.Context
 import com.auth0.android.result.Credentials
 
-object CredentialsManager {
+actual object CredentialsManager {
     private var context: Context? = null
     private val PREFERENCES_NAME = "auth0"
     private val ACCESS_TOKEN = "access_token"
@@ -12,16 +12,16 @@ object CredentialsManager {
         this.context = context
     }
 
-    fun saveCredentials(context: Context, credentials: Credentials) {
-        val sp = context.getSharedPreferences(
+    fun saveCredentials(credentials: Credentials) {
+        val sp = context?.getSharedPreferences(
             PREFERENCES_NAME, Context.MODE_PRIVATE)
 
         sp!!.edit().putString(ACCESS_TOKEN, credentials.accessToken)
             .apply()
     }
 
-    fun getAccessToken(context: Context): String? {
-        val sp = context.getSharedPreferences(
+    fun getAccessToken(): String? {
+        val sp = context?.getSharedPreferences(
             PREFERENCES_NAME, Context.MODE_PRIVATE)
 
         return sp!!.getString(ACCESS_TOKEN, null)
