+++ tutorial/shared/src/commonMain/kotlin/com/example/tutorial/shared/auth/CredentialsManager.kt
@@ -1,0 +1,6 @@
+package com.example.tutorial.shared.auth
+
+
+expect object CredentialsManager {
+    // TODO
+}
