+++ tutorial/shared/src/iosMain/kotlin/com/example/tutorial/shared/auth/CredentialsManager.kt
@@ -1,0 +1,5 @@
+package com.example.tutorial.shared.auth
+
+actual object CredentialsManager {
+    // TODO
+}
